`fn dav1d_decode_frame_init_cdf`: Remove redundant `as` casts.

+++ src/decode.rs
@@ -5649,13 +5649,13 @@ pub unsafe extern "C" fn dav1d_decode_frame_init(f: *mut Dav1dFrameContext) -> l
 #[no_mangle]
 pub unsafe extern "C" fn dav1d_decode_frame_init_cdf(f: *mut Dav1dFrameContext) -> libc::c_int {
     let c: *const Dav1dContext = (*f).c;
-    let mut retval: libc::c_int = -(22 as libc::c_int);
+    let mut retval: libc::c_int = -22;
     if (*(*f).frame_hdr).refresh_context != 0 {
         dav1d_cdf_thread_copy((*f).out_cdf.data.cdf, &mut (*f).in_cdf);
     }
     let mut tile_row = 0;
     let mut tile_col = 0;
-    (*f).task_thread.update_set = 0 as libc::c_int;
+    (*f).task_thread.update_set = 0;
     let mut i = 0;
     while i < (*f).n_tile_data {
         let mut data: *const uint8_t = (*((*f).tile).offset(i as isize)).data.data;
@@ -5669,8 +5669,8 @@ pub unsafe extern "C" fn dav1d_decode_frame_init_cdf(f: *mut Dav1dFrameContext) 
                 if (*(*f).frame_hdr).tiling.n_bytes as size_t > size {
                     return retval;
                 }
-                tile_sz = 0 as libc::c_int as size_t;
-                let mut k: libc::c_uint = 0 as libc::c_int as libc::c_uint;
+                tile_sz = 0;
+                let mut k: libc::c_uint = 0;
                 while k < (*(*f).frame_hdr).tiling.n_bytes {
                     let fresh37 = data;
                     data = data.offset(1);
@@ -5678,9 +5678,7 @@ pub unsafe extern "C" fn dav1d_decode_frame_init_cdf(f: *mut Dav1dFrameContext) 
                     k = k.wrapping_add(1);
                 }
                 tile_sz = tile_sz.wrapping_add(1);
-                size = (size as libc::c_ulong)
-                    .wrapping_sub((*(*f).frame_hdr).tiling.n_bytes as libc::c_ulong)
-                    as size_t as size_t;
+                size = size.wrapping_sub((*(*f).frame_hdr).tiling.n_bytes as usize);
                 if tile_sz > size {
                     return retval;
                 }
@@ -5694,43 +5692,42 @@ pub unsafe extern "C" fn dav1d_decode_frame_init_cdf(f: *mut Dav1dFrameContext) 
                 tile_sz,
                 tile_row as usize,
                 fresh38 as usize,
-                if (*c).n_fc > 1 as libc::c_uint {
+                if (*c).n_fc > 1 {
                     *((*f).frame_thread.tile_start_off).offset(j as isize) as usize
                 } else {
                     0
                 },
             );
             if tile_col == (*(*f).frame_hdr).tiling.cols {
-                tile_col = 0 as libc::c_int;
+                tile_col = 0;
                 tile_row += 1;
             }
             if j == (*(*f).frame_hdr).tiling.update && (*(*f).frame_hdr).refresh_context != 0 {
-                (*f).task_thread.update_set = 1 as libc::c_int;
+                (*f).task_thread.update_set = 1;
             }
             data = data.offset(tile_sz as isize);
-            size = (size as size_t).wrapping_sub(tile_sz) as size_t as size_t;
+            size = size.wrapping_sub(tile_sz);
             j += 1;
         }
         i += 1;
     }
-    if (*c).n_tc > 1 as libc::c_uint {
-        let uses_2pass: libc::c_int = ((*c).n_fc > 1 as libc::c_uint) as libc::c_int;
+    if (*c).n_tc > 1 {
+        let uses_2pass: libc::c_int = ((*c).n_fc > 1) as libc::c_int;
         let mut n = 0;
         while n < (*f).sb128w * (*(*f).frame_hdr).tiling.rows * (1 + uses_2pass) {
             reset_context(
                 &mut *((*f).a).offset(n as isize),
                 (*(*f).frame_hdr).frame_type & 1 == 0,
                 if uses_2pass != 0 {
-                    1 as libc::c_int
-                        + (n >= (*f).sb128w * (*(*f).frame_hdr).tiling.rows) as libc::c_int
+                    1 + (n >= (*f).sb128w * (*(*f).frame_hdr).tiling.rows) as libc::c_int
                 } else {
-                    0 as libc::c_int
+                    0
                 },
             );
             n += 1;
         }
     }
-    retval = 0 as libc::c_int;
+    retval = 0;
     return retval;
 }
